refactor vehicles/users package / vehichle status enum added seller class
