+++ src/main/java/com/develhope/spring/users/Seller.java
@@ -1,0 +1,4 @@
+package com.develhope.spring.users;
+
+public class Seller extends User {
+}
