+++ src/main/java/com/develhope/spring/vehicle/Car.java
@@ -1,0 +1,4 @@
+package com.develhope.spring.vehicle;
+
+public class Car extends Vehicle {
+}
