+++ src/main/java/com/develhope/spring/vehicle/Scooter.java
@@ -1,4 +1,4 @@
-package com.develhope.spring.vehicles;
+package com.develhope.spring.vehicle;
 
 public class Scooter extends Vehicle {
 }
