+++ src/main/java/com/develhope/spring/vehicle/Van.java
@@ -1,0 +1,4 @@
+package com.develhope.spring.vehicle;
+
+public class Van extends Vehicle {
+}
