--- src/main/java/com/develhope/spring/vehicles/Car.java
@@ -1,4 +1,0 @@
-package com.develhope.spring.vehicles;
-
-public class Car extends Vehicle {
-}
