--- src/main/java/com/develhope/spring/vehicles/Van.java
@@ -1,4 +1,0 @@
-package com.develhope.spring.vehicles;
-
-public class Van extends Vehicle {
-}
