--- src/main/java/com/develhope/spring/vehicles/vehicleStatus.java
@@ -1,8 +1,0 @@
-package com.develhope.spring.vehicles;
-
-public enum vehicleStatus {
-
-    UNAVAILABLE,
-    CAN_BE_ORDER,
-    PROMPT_DELIVERY,
-}
